requirements.txt
opencv-python
numpy
# flask
connexion[swagger-ui]
flask_sqlalchemy # to be removed
flask_marshmallow # to be removed
marshmallow_sqlalchemy
# sqlalchemy>=
uuid
jwt
werkzeug
flask_cors # to be removed
pycocotoolse
tqdm
tifffile
imagecodecs
